adds player name UI, attempts to create game end function
The game end function attempts to check the text content of each square and compare it to a matching array of the winning space
combinations. I have been unable to find a way to get the text content of each div when they are the value pairs of an array.

The solution I have in mind is to scrap the gameBoard object creator and send the div.space items directly into an array, to remove the issue of iterating through several objects and then attempting to target the textContent of each value.
I have avoided  using class names because I didn't want to remove the ability to generate a larger board in the future. The solution to this may be changing the class of each element on click as well as the text content.
That way the programs still maintains flexibility.