fixes game end logic
By adjusting the value of each 'this' in the game board object I am able to log each button press.
Then by adding the sum of the items within winning board the function decides a winner.

The next step in the logic is figuring out how to end the game in a tie.